move ncempy.neval.stack_align test
